Fix NumberWithUnitFormInputComponent not restoring its value after un-toggling "unlimited"
NumberWithUnitFormInputComponent has a feature which restores the input value once the "unlimited" checkbox is turned off. However, that feature seems to have regressed in some of the past refactorings. This commit fixes the restore feature and adds a test to prevent future regressions.

A note on replacing `lastRealValue` with `lastTextValue`: older implementations (before a couple refactorings) were calling `.setValue` on a nested `vcd-form-input` and that is why we needed the computed real value. Currently that is no longer the case, and instead, the `getValue` method computes the real value from the text input. Hence replacing `lastRealValue` with `lastTextValue` and restoring it instead.

Signed-off-by: Davi Barreto <dbarreto@vmware.com>